requirements.txt
awscli==1.11.13
beautifulsoup4==4.4.1
botocore==1.4.70
chardet==2.3.0
colorama==0.3.7
command-not-found==0.3
defer==1.0.6
docutils==0.12
html5lib==0.999
jmespath==0.9.0
language-selector==0.1
lxml==3.5.0
nltk==3.2.2
numpy==1.12.0
pbr==1.10.0
Pillow==3.1.2
protobuf==3.2.0
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.4.2
python-debian==0.1.27
python-systemd==231
pyxdg==0.25
requests==2.9.1
roman==2.0.0
rsa==3.2.3
s3transfer==0.1.9
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
sklearn==0.0
ssh-import-id==5.5
stevedore==1.16.0
system-service==0.3
tensorflow-gpu==1.0.1
tqdm==4.11.2
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
